requirements.txt
python-dotenv
cdktf-cdktf-provider-aws